Fix Exploit regarding Jewel Enhancement
Fix Exploit: When the item name is longer than 30 chars, the Jewel Enhancement level could not be written, leading to infinite enhancements - an Easter egg is implemented to change the item's name to stop this.
Fix(?) Oversight: sing infinite stamina items as material will not consume the item.
Add: Glitch text effects to use for club21's lore material.