Update titles_1.php
To deal with player dynamic is an endless struggle. We're supposed to only build mechanics and aesthetics for players, and dynamic is not one of them. The player must carve out their own dynamic, and only with that, we adjust based on what the player would say.
In this case, something once deemed too strong is in fact underused by the player based on their feedback, so from the same feedback, this certain stat for this certain item is changed to reflect a dynamic not-in-the-design.